Update README.md
Ansible is used in this assignment to install a Python program and HAProxy on three distinct servers. Three web servers are used in the deployment along with a load balancer running HAProxy, and another server serves as a Bastion host for SSH access to all other hosts.

The site. yaml file in the Ansible playbook (site. yaml) assumes that all required servers have previously been set up and that SSH keys are set up for password-free SSH access between the Bastion host and the other hosts. Using an SSH configuration file that permits the host to utilize the Bastion host as a jump host, the playbook may be executed from outside the site but through the Bastion host.

User names, SSH key paths, python application settings, and other variables may be added or changed as needed on the site.yaml playbook.